Moved InteractionAwareActivity to 'common' module in order to be able to use it in other submodules (ie uphold-integration). Fixed crash when manipulating UI elements after finishing Activity (in uphold-integration).
