+++ wallet/src/de/schildbach/wallet/ui/BaseMenuActivity.kt
@@ -21,11 +21,11 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.MenuItem
 import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import de.schildbach.wallet.WalletApplication
 import de.schildbach.wallet_test.R
 import org.dash.wallet.common.Configuration
+import org.dash.wallet.common.InteractionAwareActivity
 
 /**
  * @author Samuel Barbosa
