+++ wallet/src/de/schildbach/wallet/ui/UpholdTransferActivity.kt
@@ -34,6 +34,7 @@ import de.schildbach.wallet.ui.send.EnterAmountSharedViewModel
 import de.schildbach.wallet_test.R
 import org.bitcoinj.core.Coin
 import org.bitcoinj.utils.MonetaryFormat
+import org.dash.wallet.common.InteractionAwareActivity
 import org.dash.wallet.common.util.GenericUtils
 import org.dash.wallet.integration.uphold.data.UpholdTransaction
 import org.dash.wallet.integration.uphold.ui.UpholdWithdrawalHelper
